Replace rakefile with rakefile.netstd
