+++ build.cmd
@@ -1,3 +1,3 @@
 @echo off
 
-call bundle exec rake %*
+call bundle exec rake -f rakefile.netstd.rb %*
