--- build.netstd.cmd
@@ -1,3 +1,0 @@
-@echo off
-
-call bundle exec rake -f rakefile.netstd.rb %*
